Skeleton Layout for the App
I made some changes to the navigation that uses fragments instead of activities. I set up most of the possible pages we need for the app except for a few (how the session is going to work). The layout is temporary.

Message me on discord if something doesn't makes sense or doesn't work on your end.